+++ src/ivory/utils/config_section.py
@@ -1,0 +1,9 @@
+from ivory.utils.classproperty import classproperty
+
+
+class ConfigSection(dict):
+    """Class to represent the config file sections."""
+
+    @classproperty
+    def name(cls):
+        return cls.__name__