+++ tests/config/workflow_config.py
@@ -2,7 +2,7 @@ from ivory.loop import Loop
 from ivory.utils.config_section import ConfigSection
 
 Pipeline = ConfigSection(
-    plugins=Loop(["test.plugin.simple_plugin", "test.plugin.simple_plugin"])
+    plugins=Loop(["tests.plugin.simple_plugin", "tests.plugin.simple_plugin"])
 )
 
 MockPlugin = ConfigSection(