+++ tests/config/workflow_config_load_context.py
@@ -4,7 +4,7 @@ from ivory.loop import Loop
 from ivory.utils.config_section import ConfigSection
 
 Pipeline = ConfigSection(
-    plugins=Loop(["test.plugin.simple_plugin"]),
+    plugins=Loop(["tests.plugin.simple_plugin"]),
     context=os.path.join(os.getcwd(), "cache/simple_plugin.pickle"),
 )