+++ tests/config/workflow_config_simple.py
@@ -1,5 +1,5 @@
 from ivory.utils.config_section import ConfigSection
 
 Pipeline = ConfigSection(
-    plugins=["test.plugin.simple_plugin", "test.plugin.simple_plugin"]
+    plugins=["tests.plugin.simple_plugin", "tests.plugin.simple_plugin"]
 )